termino das telas iniciais de admin e ajuste de estilos
