+++ front/src/components/admin/categoriasGerenciamento/categoriasGerenciamento.css
@@ -1,0 +1,15 @@
+#button{
+    width: 70%;
+    display: flex;
+    align-items: flex-end;
+    justify-content: end;
+}
+
+.bi-trash {
+    color: red;
+    margin-right: 0.3rem;
+}
+
+#excluir{
+    color: red;
+}
