+++ front/src/components/admin/globalAdmin.css
@@ -1,0 +1,232 @@
+@import url("https://cdn.jsdelivr.net/npm/bootstrap-icons@1.13.1/font/bootstrap-icons.min.css");
+
+.header-admin{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin: 1.5rem 0;
+}
+
+.header-admin p, h1{
+    margin: 0.5rem;
+}
+
+.bodyAdmin{
+    padding: 1rem 1rem 1rem 0;
+}
+
+.fundo-container-admin {
+  display: flex;
+  justify-content: center;
+  background-color: rgb(81, 108, 202,0.10);
+  height: 95vh;
+  width: 80%;
+  box-sizing: border-box;
+  border-radius: 8px;
+  align-content: center;
+  margin-top: 1rem;
+}
+
+.fundo-content-admin {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  justify-items: center;
+
+}
+
+.navAdmin{
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    height: 95vh;
+    width: 18rem;
+    padding: 1rem 0;
+    align-items: center;
+    box-sizing: border-box;
+}
+
+.mainAdmin{
+    display: flex;
+}
+
+.links-nav{
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+}
+
+.links-nav a{
+    text-decoration: none;
+    color: #3498DB;
+    border: 1px solid #3498DB;
+    width: 7rem;
+    padding: 0.8rem 1.9rem;
+    border-radius: 8px;
+    display: flex;
+    justify-content: flex-start;
+    gap: 0.5rem;
+}
+
+.links-nav a:hover{
+    background-color: #3498DB;
+    color: white;
+}
+
+.fundo-content-admin h1 {
+  color: #3498DB;
+}
+
+.fundo-content-admin p{
+ color: rgb(141, 137, 137, 0.59); 
+}
+
+.table-admin {
+    margin: 1rem 0;
+    width: 70%;
+    background-color: #ffffff;
+    padding: 25px;
+    border-radius: 10px;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+    box-sizing: border-box;
+}
+/* Cabeçalho: Pesquisa e Filtro */
+.search-field {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 20px;
+}
+
+#search-input {
+    flex-grow: 1;
+    padding: 10px 15px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    font-size: 16px;
+    margin-right: 15px;
+}
+
+.filter-btn {
+    padding: 10px 20px;
+    background-color: #ffffff;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    font-size: 16px;
+    cursor: pointer;
+    font-weight: bold;
+    color: #555;
+}
+
+/* Tabela */
+
+.user-table {
+    width: 100%;
+    border-collapse: collapse;
+    table-layout: fixed;
+}
+
+.user-table th, .user-table td {
+    padding: 15px;
+    text-align: left;
+    border-bottom: 1px solid #e0e0e0;
+}
+
+.user-table thead {
+    background-color: #3498DB;
+    color: white;
+}
+.user-table tbody {
+    max-width: 100%;
+}
+
+.user-table th {
+    font-weight: bold;
+}
+
+.user-table th.sortable {
+    cursor: pointer;
+}
+
+/* Estilos para Status (Ativo/Inativo) */
+.status {
+    padding: 5px 12px;
+    border-radius: 15px;
+    font-size: 12px;
+    font-weight: bold;
+    color: #fff;
+    display: inline-block;
+    text-align: center;
+    min-width: 60px;
+}
+
+.status-ativo {
+    background-color: #28a745; /* Verde */
+}
+
+.status-inativo {
+    background-color: #dc3545; /* Vermelho */
+}
+
+/* Estilos para os Botões de Tipo */
+.btn-tipo {
+    padding: 8px 15px;
+    border-radius: 5px;
+    cursor: pointer;
+    font-size: 14px;
+    font-weight: bold;
+    background-color: transparent;
+}
+
+.questionador {
+    border: 1px solid #3498DB;
+    color: #3498DB;
+}
+
+.respondente {
+    border: 1px solid #3498DB;
+    color: #3498DB;
+}
+
+/* Rodapé e Paginação */
+.table-footer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding-top: 20px;
+}
+
+.pagination a {
+    color: #555;
+    text-decoration: none;
+    padding: 5px 5px;
+    margin: 0 4px;
+    border-radius: 4px;
+}
+
+.pagination a.active {
+    color: #3C509A;
+    border-color: #3C509A;
+}
+
+.pagination a:hover:not(.active) {
+    background-color: #f1f1f1;
+}
+
+#cadastrar-categoria-btn{
+    text-decoration: none;
+    color: #3498DB;
+    border: 1px solid #3498DB;
+    padding: 0.6rem 0.8rem;
+    border-radius: 5px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 0.5rem;
+    background-color: white;
+}
+
+#cadastrar-categoria-btn:hover{
+    background-color: #3498DB;
+    color: white;
+}
