+++ front/src/components/user/SobreNos/SobreNos.css
@@ -1,0 +1,78 @@
+.sobre-nos {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh; /* altura mínima: 100% da viewport */
+}
+
+.sobre-nos-introducao {
+  max-width: 800px;
+  margin: auto;
+  text-align: center;
+  margin-bottom: 3rem;
+}
+
+.sobre-nos-introducao h2 {
+  font-size: 1.5rem;
+  color: #004a99;
+  margin-bottom: 1rem;
+}
+
+.sobre-nos-introducao p {
+  font-size: 1.1rem;
+  line-height: 1.6;
+}
+
+.sobre-nos-integrantes {
+  text-align: center;
+}
+
+.sobre-nos-integrantes h3 {
+  font-size: 1.7rem;
+  margin-bottom: 2rem;
+  color: #004a99;
+}
+.integrantes-circulares {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: 2rem;
+  margin-top: 2rem;
+}
+
+.integrante-circular {
+  text-align: center;
+  max-width: 180px;
+}
+
+.foto-circular {
+  width: 120px;
+  height: 120px;
+  border-radius: 50%;
+  object-fit: cover;
+  border: 4px solid #fff;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+  margin-bottom: 0.8rem;
+}
+
+.integrante-circular h4 {
+  font-size: 1rem;
+  color: #003366;
+  margin-bottom: 0.3rem;
+}
+
+.integrante-circular p {
+  font-size: 0.9rem;
+  color: #333;
+  line-height: 1.3;
+}
+
+.footer-scroll {
+  position: relative;
+  width: 100%;
+  height: 3.5rem;
+  background-color: var(--azul-escuro);
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  margin-top: auto; /* ESSENCIAL para empurrar o rodapé ao fim da página */
+}
