+++ front/src/components/user/global.css
@@ -1,0 +1,172 @@
+.fundo-gradiente{
+  background: var(--gradiente-fundo);
+  min-height: 100vh;
+}
+
+/* Cabeçalho Global */
+.header-global {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  border-bottom: 0px;
+  width: 100%;
+  background: var(--gradiente-fundo);
+}
+
+.header-global-nav {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  padding: 10px 20px;
+}
+
+.nav-bar i {
+  font-size: 1.2rem;
+  margin-right: 3px;
+}
+
+.nav-bar-item {
+  color: var(--azul-principal);
+  margin-left: 2%;
+  font-size: 0.93rem;
+  font-weight: 700;
+  margin-top: 1.5rem;
+}
+
+/* Main content area */
+.titulo-pagina {
+  position: absolute;
+  left: 50%;
+  transform: translateX(-50%);
+  white-space: nowrap;
+  color: var(--azul-principal);
+}
+
+.painel-titulo {
+  font-size: 1.8rem;
+  color: var(--azul-principal);
+}
+
+.cards {
+  display: flex;
+  justify-content: center;
+  gap: 20px;
+  flex-wrap: wrap;
+  margin-bottom: 40px;
+  align-items: center;
+}
+
+.card {
+  width: 15rem;
+  height: 15rem;
+  background: var(--gradiente-card);
+  border-radius: 4px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  text-decoration: none;
+  color: white;
+  transition: background 0.3s;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.card:hover {
+  background: var(--gradiente-card-hover);
+}
+
+.card-img {
+  width: 50px;
+  height: 50px;
+  margin-bottom: 12%;
+}
+
+.card-title {
+  font-size: 1.3rem;
+  font-weight: 500;
+  text-align: center;
+}
+
+footer img {
+  height: 3rem;
+  object-fit: contain;
+  margin: 1%;
+}
+
+.filtrar-button {
+  color: #3498db;
+  width: auto;
+  padding: 8px 12px;
+  background-color: #fff;
+  border: 1px solid #3498db;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: background-color 0.2s ease;
+  margin-top: 0;
+}
+
+.filtrar-button:hover {
+  background-color: #4ea1fb29;
+}
+
+.logout-button {
+  background-color: transparent;
+  border: 2px solid red;
+  color: red;
+  font-size: 1.1rem;
+  font-weight: 600;
+  border-radius: 6px;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  white-space: nowrap;
+  width: 11rem;
+  height: 3.15rem;
+  box-sizing: border-box;
+  text-align: center;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.logout-button:hover {
+  background-color: red;
+  color: white;
+}
+
+
+@media (max-width: 768px) {
+  .nav-bar-item {
+    font-size: 2.25vw;
+  }
+
+  .pagina-titulo {
+    font-size: 1.5rem;
+  }
+
+  .cards {
+    flex-direction: column;
+    align-items: center;
+    gap: 5vw;
+  }
+
+  .card {
+    width: 50%;
+    aspect-ratio: 1 / 1;
+  }
+
+  .card-img {
+    width: 15%;
+    height: auto;
+    margin-bottom: 8%;
+  }
+
+  .card-title {
+    font-size: 4vw;
+  }
+
+  .logout-button {
+    width: 30%;
+    font-size: 4vw;
+    padding: 1.5% 0;
+  }
+}
